requirements.txt
asgiref==3.8.1
Django==5.0.4
django-filter==24.2
sqlparse==0.4.4
tzdata==2024.1
